Git-blame: Diff only this file
In large commit-diffs it is hard to find the file you are working on
and there seems to be quite a lot of projects that have huge commits.
If your commit is large enough, you might crash Kate because you run out
of memory.

This change addresses both these problems by only showing the diff for
the current file.

BUG: 443387